+++ packages/core/test/wrapper.test.ts
@@ -281,6 +281,37 @@ describe('Tests Native Wrapper', () => {
       expect(RNSentry.setContext).not.toBeCalled();
       expect(RNSentry.setExtra).not.toBeCalled();
     });
+
+    test('sets ignoreErrorsStr and ignoreErrorsRegex correctly when ignoreErrors contains strings and regex', async () => {
+      const regex1 = /foo/;
+      const regex2 = new RegExp('bar');
+      await NATIVE.initNativeSdk({
+        dsn: 'test',
+        enableNative: true,
+        ignoreErrors: ['string1', regex1, 'string2', regex2],
+        devServerUrl: undefined,
+        defaultSidecarUrl: undefined,
+        mobileReplayOptions: undefined,
+      });
+      expect(RNSentry.initNativeSdk).toBeCalled();
+      const initParameter = (RNSentry.initNativeSdk as jest.MockedFunction<any>).mock.calls[0][0];
+      expect(initParameter.ignoreErrorsStr).toEqual(['string1', 'string2']);
+      expect(initParameter.ignoreErrorsRegex).toEqual([regex1.source, regex2.source]);
+    });
+
+    test('does not set ignoreErrorsStr or ignoreErrorsRegex if ignoreErrors is not provided', async () => {
+      await NATIVE.initNativeSdk({
+        dsn: 'test',
+        enableNative: true,
+        devServerUrl: undefined,
+        defaultSidecarUrl: undefined,
+        mobileReplayOptions: undefined,
+      });
+      expect(RNSentry.initNativeSdk).toBeCalled();
+      const initParameter = (RNSentry.initNativeSdk as jest.MockedFunction<any>).mock.calls[0][0];
+      expect(initParameter.ignoreErrorsStr).toBeUndefined();
+      expect(initParameter.ignoreErrorsRegex).toBeUndefined();
+    });
   });
 
   describe('sendEnvelope', () => {