+++ packages/core/android/src/main/java/io/sentry/react/RNSentryModuleImpl.java
@@ -83,6 +83,7 @@ import java.io.InputStream;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.nio.charset.Charset;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
@@ -90,6 +91,7 @@ import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
 import java.util.concurrent.CountDownLatch;
+import java.util.regex.Pattern;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.TestOnly;
@@ -319,6 +321,8 @@ public class RNSentryModuleImpl {
     // React native internally throws a JavascriptException.
     // we want to ignore it on the native side to avoid sending it twice.
     options.addIgnoredExceptionForType(JavascriptException.class);
+
+    trySetIgnoreErrors(options, rnOptions);
 
     options.setBeforeSend(
         (event, hint) -> {
@@ -1125,5 +1129,37 @@ public class RNSentryModuleImpl {
       return null;
     }
     return uri.getScheme() + "://" + uri.getHost();
+  }
+
+  @TestOnly
+  protected void trySetIgnoreErrors(SentryAndroidOptions options, ReadableMap rnOptions) {
+    ReadableArray regErrors = null;
+    ReadableArray strErrors = null;
+    if (rnOptions.hasKey("ignoreErrorsRegex")) {
+      regErrors = rnOptions.getArray("ignoreErrorsRegex");
+    }
+    if (rnOptions.hasKey("ignoreErrorsStr")) {
+      strErrors = rnOptions.getArray("ignoreErrorsStr");
+    }
+    if (regErrors == null && strErrors == null) {
+      return;
+    }
+
+    int regSize = regErrors != null ? regErrors.size() : 0;
+    int strSize = strErrors != null ? strErrors.size() : 0;
+    List<String> list = new ArrayList<>(regSize + strSize);
+    if (regErrors != null) {
+      for (int i = 0; i < regErrors.size(); i++) {
+        list.add(regErrors.getString(i));
+      }
+    }
+    if (strErrors != null) {
+      // Use the same behaviour of JavaScript instead of Android when dealing with strings.
+      for (int i = 0; i < strErrors.size(); i++) {
+        String pattern = ".*" + Pattern.quote(strErrors.getString(i)) + ".*";
+        list.add(pattern);
+      }
+    }
+    options.setIgnoredErrors(list);
   }
 }