+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ## Unreleased
 
-- ignoreError now filters Native errors ([#4930](https://github.com/getsentry/sentry-react-native/pull/4930))
+- ignoreError now filters Native errors ([#4948](https://github.com/getsentry/sentry-react-native/pull/4948))
 
 you can use strings to filter errors that contains your string or RegEx for filtering for a pattern.