+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 - ignoreError now filters Native errors ([#4948](https://github.com/getsentry/sentry-react-native/pull/4948))
 
-you can use strings to filter errors that contains your string or RegEx for filtering for a pattern.
+You can use strings to filter errors or RegEx for filtering with a pattern.
 
 example: